Move some hash types to primitives
In preparation for moving the `block` module to `primitives` move the
transaction id types and Merkle tree hash types to `primitives`.
